Don't check password on upload
It's not strictly necessary to check the password when uploading a file. Before cli uploads was supported by a module for nginx that is not supported since several years. But not using it made cli uploads to password protected tickets impossible.

I don't think this will cause any problems for the service, you still need a password to see and download files.